+++ singletons-base/tests/compile-and-dump/Singletons/T172.hs
@@ -13,6 +13,6 @@ import GHC.TypeLits.Singletons
 import Prelude.Singletons
 
 $(singletonsOnly [d|
-  ($>) :: Nat -> Nat -> Nat
+  ($>) :: Natural -> Natural -> Natural
   ($>) = (+)
   |])